Handle e-mail change
Some users change their e-mail address while the action is ongoing.
Since this is used internally as a key it was a problem for graphs which
had duplicate names.

The graphs now include name and e-mail address which makes the entries
unique again.

Also increased the graph size since the legend takes more room now.